Added Validator and Exception

+++ service/src/main/java/com/service/impl/RestaurantServiceImpl.java
@@ -5,6 +5,9 @@ import com.service.RestaurantService;
 import com.domain.entity.RestaurantEntity;
 import com.domain.model.Restaurant;
 import com.service.converter.Converter;
+
+import com.service.validator.Validator;
+
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,6 +29,8 @@ public class RestaurantServiceImpl implements RestaurantService {
 
     @Override
     public Restaurant getById(long id) {
+    	Validator.checkId(id);
+    	Validator.checkEntity(restaurantRepository.findRestaurantEntityById(id));
         return converter.entityToRestaurant(restaurantRepository.findRestaurantEntityById(id));
     }
 
@@ -36,14 +41,18 @@ public class RestaurantServiceImpl implements RestaurantService {
 
     @Override
     public Restaurant save(Restaurant restaurant) {
+    	Validator.checkEntity(restaurant);
         restaurantRepository.save(converter.restaurantToEntity(restaurant));
         return restaurant;
     }
 
     @Override
-    public void update(long id, Restaurant restaurant) {
-        RestaurantEntity restaurantEntity = restaurantRepository.findRestaurantEntityById(id);
-        if(restaurant != null) {
+    public void update(long id,Restaurant restaurant) {
+    	Validator.checkId(id);
+    	RestaurantEntity restaurantEntity = restaurantRepository.findRestaurantEntityById(id);
+        Validator.checkEntity(restaurant);
+        Validator.checkEntity(restaurantEntity);
+   
         	if(restaurant.getName() != null) {
         		restaurantEntity.setName(restaurant.getName());
         	}
@@ -67,13 +76,16 @@ public class RestaurantServiceImpl implements RestaurantService {
         	}
         	if(restaurant.getPhoneNumber() != null) {
         		restaurantEntity.setPhoneNumber(restaurant.getPhoneNumber());
+
         	}
-    }
         restaurantRepository.save(restaurantEntity);
     }
 
     @Override
     public void delete(long id) {
-        restaurantRepository.deleteById(id);
+    	Validator.checkId(id);
+    	if(Validator.checkEntity(restaurantRepository.findRestaurantEntityById(id))) {
+    		 restaurantRepository.deleteById(id);
+    	}
     }
 }
